move new group split tests to test_connections.py
